+++ packages/image_picker/image_picker_platform_interface/lib/src/platform_interface/image_picker_platform.dart
@@ -167,21 +167,4 @@ abstract class ImagePickerPlatform extends PlatformInterface {
   }) {
     throw UnimplementedError('pickVideo() has not been implemented.');
   }
-
-  /// Retrieve the lost [PickedFile] file when [pickImage] or [pickVideo] failed because the MainActivity is destroyed. (Android only)
-  ///
-  /// Image or video can be lost if the MainActivity is destroyed. And there is no guarantee that the MainActivity is always alive.
-  /// Call this method to retrieve the lost data and process the data according to your APP's business logic.
-  ///
-  /// Returns a [LostData] object if successfully retrieved the lost data. The [LostData] object can represent either a
-  /// successful image/video selection, or a failure.
-  ///
-  /// Calling this on a non-Android platform will throw [UnimplementedError] exception.
-  ///
-  /// See also:
-  /// * [LostData], for what's included in the response.
-  /// * [Android Activity Lifecycle](https://developer.android.com/reference/android/app/Activity.html), for more information on MainActivity destruction.
-  Future<LostData> retrieveLostData() {
-    throw UnimplementedError('retrieveLostData() has not been implemented.');
-  }
 }