+++ packages/image_picker/image_picker_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+
+* Remove method `retrieveLostDataAsDartIoFile()` eliminating `dart.io` library dependency for class `ImagePickerPlatform`. 
+
 ## 1.1.0
 
 * Introduce PickedFile type for the new API.